Radar implementation: Add Mock 2 and On/Off buttons
Adding a file, "mock2.py" is another example of data from lidar.
Changed the "Example" button to "Enable". Launches the mock and modifies a variable to activate the radar.
The "Deactivate" button modifies this variable so that the mock no longer runs.
The "clear" button clears all cache data and displays the radar without points.