**Fix : CORS Issue with Image Display in MapList**
After addressing the reported issue where saved maps didn't appear in the MapList and resulted in empty images, a fix has been implemented.
Initially, there was an attempt to modify the way image links from Amazon S3 were accessed.
Transitioning from using navigation props to creating an API route to retrieve map links from the backend.

Upon further testing, it was observed that the CORS policy issue persisted, particularly with older maps stored in the database.
To fully resolve this, it's recommended to follow these steps:

1. **Delete Existing Maps:** Remove all existing maps associated with your user profile.
2. **Add New Maps:** After clearing the old maps, add new maps to your profile.

This process ensures that the CORS-related problem is resolved for both existing and new maps.